Detective UI (Clue Solver)
Built the basics of the Detective UI in the Hideout. The player can use the UI to select relevant options to solve the current event's parameters (location, gang, attack type).
Currently, the UI updates to reflect the player's selections via the Detective UI's images and text options, but does not actually match with the ClueMaster's event parameters.
The next step is to reconcile that. As such, I NEED the "selections" and "solutions" to match up perfectly.